requirements.txt
Django>=4.0
djangorestframework
psycopg2-binary
celery
redis
gunicorn
openpyxl
python-decouple  # for loading .env files
psycopg2-binary
dj-database-url  # for parsing the DATABASE_URL
django-cors-headers  # for handling CORS headers
celery
djangorestframework
redis
